Update index.md
For some reason, when I pulled this section up in Visual Studio half of it was missing as well as the lecture section. @gstark 
I did not see any updates that needed to be done in the lecture, except that in slide 12 the bottom is cut off. So it stops in mid sentence